+++ CHANGELOG.md
@@ -8,7 +8,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 ### Added
 * **BREAKING** Pagination: Navigatie element toevoegen ([#2758](https://github.com/dso-toolkit/dso-toolkit/issues/2758))
 * Tabs: Web Component implementatie ([#2236](https://github.com/dso-toolkit/dso-toolkit/issues/2236))
-* Projects Item: Nieuw component ([#2779](https://github.com/dso-toolkit/dso-toolkit/issues/2779))
+* Project Item: Nieuw component ([#2779](https://github.com/dso-toolkit/dso-toolkit/issues/2779))
 
 ### Changed
 * Document Component: Panel heading wordt "Kenmerken en kaart" ([#2798](https://github.com/dso-toolkit/dso-toolkit/issues/2798))